VideoPosts Correctly displaying valid URLS only
