+++ edtech-social-app/package.json
@@ -27,7 +27,9 @@
     "postcss": "^8",
     "prettier": "^3.3.3",
     "prettier-plugin-tailwindcss": "^0.6.5",
+    "react-player": "^2.16.0",
     "tailwindcss": "^3.4.1",
-    "typescript": "^5"
+    "typescript": "^5",
+    "validator": "^13.12.0"
   }
 }
