+++ edtech-social-app/utils/utils.ts
@@ -1,0 +1,9 @@
+// edtech-social-app\utils\utils.ts
+export const isValidVideoUrl = async (url: string): Promise<boolean> => {
+    try {
+      const response = await fetch(url, { method: 'HEAD' });
+      return response.ok;
+    } catch (error) {
+      return false;
+    }
+  };
